:recycle: refactor: Refatorando Entity Posts
